Use built-in hash function.
Change the unordered_set to save both the data pointer as well as the
length. This allows us to use gnu's built-in hash function which takes a
data pointer and length instead of using our own hash function. The
built-in hash function is a bit faster.

Also clean up the use of add_filter_value() by removing the optional
index argument. There isn't any reason why the number of filter values
needs to be maintained by the caller of add_filter_value(), and removing
the index makes sure it's not possible to add multiple values at the
same index, which would break an assumption of the unordered set.