Take advantage of string length.
Take advantage of string length when doing set membership tests:

 - The unordered_set now holds pairs of (string,length) instead of just
   a null-terminated string. That allows using the built-in GCC hash
   funtion, which is quite fast.
 - When comparing strings in the hash equality function, only bother
   doing the buffer comparison when the string lengths match.
 - It can be inefficient to hash a very long string when all members of
   the set are short strings. To make this case faster, keep track of
   the minumum and maximum string length across the set members and only
   bother doing the set comparison of the lengths overlap.

To effectively use this, the length needs to be filled in when
sinsp_filter_check::flt_compare() is called, so do a pass over the
existing filterchecks and actually return the actual string length when
it's easily known. In flt_compare(), if the type is a string and the
provided length is 0, do a strlen() to find it. This should be very rare
now that the length is properly passed back.

I was wrong about not having to keep track of the index in
add_filter_value(), so add that code back in.

At this point, there isn't any requirement that the items in the set are
null terminated strings, so change the names of the hash/comparison
functions to use "mem" instead of "str".